outdated/flask_example/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [
    { include = "backend" },  
    { include = "ml_model" },
]

[tool.poetry.dependencies]
python = "^3.9"
asgiref = "3.4.1"
astroid = "2.9.2"
Django = "4.0.1"
django-cors-headers = "3.10.1"
django-rest-framework = "^0.1.0"
djangorestframework = "3.13.1"
flake8 = "4.0.1"
isort = "5.10.1"
lazy-object-proxy = "1.7.1"
mccabe = "0.6.1"
mysqlclient = "2.1.0"
platformdirs = "2.4.1"
pycodestyle = "2.8.0"
pyflakes = "2.4.0"
pylint = "2.12.2"
pytz = "2021.3"
sqlparse = "^0.4.2"
toml = "0.10.2"
typing-extensions = "^4.0.1"
wrapt = "1.13.3"
prompt-toolkit = "3.0.23"
paho-mqtt = "1.6.1"
Pillow = "9.0.0"
aioredis = "1.3.1"
amqp = "5.0.9"
async-timeout = "4.0.2"
attrs = "21.4.0"
autobahn = "21.11.1"
colorama = "0.4.4"
Automat = "20.2.0"
billiard = "3.6.4.0"
celery = "5.2.3"
cffi = "1.15.0"
channels = "3.0.4"
channels-redis = "3.3.1"
click = "8.0.3"
click-didyoumean = "0.3.0"
click-plugins = "1.1.1"
click-repl = "0.2.0"
constantly = "15.1.0"
cryptography = "36.0.1"
daphne = "3.0.2"
hiredis = "2.0.0"
hyperlink = "21.0.0"
idna = "3.3"
incremental = "21.3.0"
msgpack = "1.0.3"
pyasn1 = "0.4.8"
pyasn1-modules = "0.2.8"
pycparser = "2.21"
pyOpenSSL = "21.0.0"
tzdata = "2021.5"
service-identity = "21.1.0"
six = "1.16.0"
Twisted = "21.7.0"
txaio = "21.2.1"
vine = "5.0.0"
wcwidth = "0.2.5"
"zope.interface" = "5.4.0"
websocket-client = "1.2.3"

[tool.poetry.dev-dependencies]
black = "*"
pytest = "*"
coverage = "*"
flake8 = "*"
bandit = "*"
ipykernel = "*"
yapf = "*"
pylint = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
